created folder structure for tables, joint tables, relations and joint relations with corresponding items
