+++ src/db/schema.ts
@@ -1,9 +1,10 @@
 import { account, session, user, verification } from "./auth-schema"
-import { tools } from "./tools";
-import { categories } from "./categories";
-import { platforms } from "./platforms";
-import { tags } from "./tags";
-import { companies } from "./companies";
-import { features } from "./features";
+import { tools } from "./tables/tools";
+import { categories } from "./tables/categories";
+import { platforms } from "./tables/platforms";
+import { tags } from "./tables/tags";
+import { companies } from "./tables/companies";
+import { features } from "./tables/features";
+import { reactions } from "./tables/reactions";
 
-export { account, session, user, verification, tools, categories, platforms, tags, companies, features }
+export { account, session, user, verification, tools, categories, platforms, tags, companies, features, reactions }
