+++ src/db/tables/categories.ts
