+++ src/db/tables/platforms.ts
