+++ src/db/tables/reactions.ts
@@ -1,0 +1,12 @@
+import { pgTable, text, uuid } from "drizzle-orm/pg-core";
+import { tools } from "./tools";
+
+// --- Reactions ---
+export const reactions = pgTable('reactions', {
+  id: uuid('id').primaryKey().defaultRandom(),
+  userId: uuid('user_id').notNull(),
+  toolId: uuid('tool_id')
+    .notNull()
+    .references(() => tools.id, { onDelete: 'cascade' }),
+  reactType: text('react_type').notNull(), // e.g., "like", "love", etc.
+});
